+++ lib/servactory/test_kit/result.rb
@@ -3,10 +3,26 @@
 module Servactory
   module TestKit
     class Result
-      def initialize(**attributes)
-        attributes.each_pair do |method_name, method_return|
-          define_singleton_method(method_name) { method_return }
+      def self.as_success(attributes = {})
+        Servactory::Result.success_for(context: new(attributes))
+      end
+
+      def self.as_failure(exception: nil)
+        Servactory::Result.failure_for(exception: exception)
+      end
+
+      def initialize(attributes = {})
+        attributes.each_pair do |name, value|
+          service_storage[:outputs].merge!({ name => value })
         end
+      end
+
+      private_class_method :new
+
+      private
+
+      def service_storage
+        @service_storage ||= { outputs: {} }
       end
     end
   end