+++ spec/examples/usual/example55_spec.rb
@@ -6,56 +6,64 @@ RSpec.describe Usual::Example55 do
 
     let(:attributes) do
       {
-        first_name: first_name,
-        middle_name: middle_name,
-        last_name: last_name
+        event_name: event_name
       }
     end
 
-    let(:first_name) { "John" }
-    let(:middle_name) { "Fitzgerald" }
-    let(:last_name) { "Kennedy" }
+    let(:event_name) { "created" }
 
     include_examples "check class info",
-                     inputs: %i[first_name last_name middle_name],
+                     inputs: %i[event_name],
                      internals: %i[],
-                     outputs: %i[full_name]
+                     outputs: %i[event]
 
     context "when the input arguments are valid" do
       describe "and the data required for work is also valid" do
         include_examples "success result class"
 
-        it "returns the expected value in `full_name`" do
-          result = perform
-
-          expect(result.full_name).to eq("John Fitzgerald Kennedy")
-        end
-
-        describe "even if `middle_name` is not specified" do
-          let(:middle_name) { nil }
-
-          it "returns the expected value in `full_name`" do
+        describe "and the value of `event_name` is passed" do
+          it "returns the expected values", :aggregate_failures do
             result = perform
 
-            expect(result.full_name).to eq("John Kennedy")
+            expect(result.event).to be_a(Usual::Example55::Event)
+            expect(result.event.id).to be_present
+            expect(result.event.event_name).to eq("created")
+          end
+        end
+
+        describe "and the value of `event_name` is not passed" do
+          let(:event_name) { nil }
+
+          it "returns the expected values", :aggregate_failures do
+            result = perform
+
+            expect(result.event).to be_a(Usual::Example55::Event)
+            expect(result.event.id).to be_present
+            expect(result.event.event_name).to be_nil
+          end
+        end
+      end
+
+      describe "but the data required for work is invalid" do
+        describe "because the value of `event_name` is wrong" do
+          let(:event_name) { "sent" }
+
+          it "returns expected error" do
+            expect { perform }.to(
+              raise_error(
+                ApplicationService::Errors::InputError,
+                "[Usual::Example55] Wrong value in `event_name`, must be one of " \
+                "`[\"created\", \"rejected\", \"approved\"]`"
+              )
+            )
           end
         end
       end
     end
 
     context "when the input arguments are invalid" do
-      context "when `first_name`" do
-        it_behaves_like "input required check", name: :first_name
-        it_behaves_like "input type check", name: :first_name, expected_type: String
-      end
-
-      context "when `middle_name`" do
-        it_behaves_like "input type check", name: :middle_name, expected_type: String
-      end
-
-      context "when `last_name`" do
-        it_behaves_like "input required check", name: :last_name
-        it_behaves_like "input type check", name: :last_name, expected_type: String
+      context "when `event_name`" do
+        it_behaves_like "input type check", name: :event_name, expected_type: String
       end
     end
   end
@@ -65,56 +73,64 @@ RSpec.describe Usual::Example55 do
 
     let(:attributes) do
       {
-        first_name: first_name,
-        middle_name: middle_name,
-        last_name: last_name
+        event_name: event_name
       }
     end
 
-    let(:first_name) { "John" }
-    let(:middle_name) { "Fitzgerald" }
-    let(:last_name) { "Kennedy" }
+    let(:event_name) { "created" }
 
     include_examples "check class info",
-                     inputs: %i[first_name last_name middle_name],
+                     inputs: %i[event_name],
                      internals: %i[],
-                     outputs: %i[full_name]
+                     outputs: %i[event]
 
     context "when the input arguments are valid" do
       describe "and the data required for work is also valid" do
         include_examples "success result class"
 
-        it "returns the expected value in `full_name`" do
-          result = perform
-
-          expect(result.full_name).to eq("John Fitzgerald Kennedy")
-        end
-
-        describe "even if `middle_name` is not specified" do
-          let(:middle_name) { nil }
-
-          it "returns the expected value in `full_name`" do
+        describe "and the value of `event_name` is passed" do
+          it "returns the expected values", :aggregate_failures do
             result = perform
 
-            expect(result.full_name).to eq("John Kennedy")
+            expect(result.event).to be_a(Usual::Example55::Event)
+            expect(result.event.id).to be_present
+            expect(result.event.event_name).to eq("created")
+          end
+        end
+
+        describe "and the value of `event_name` is not passed" do
+          let(:event_name) { nil }
+
+          it "returns the expected values", :aggregate_failures do
+            result = perform
+
+            expect(result.event).to be_a(Usual::Example55::Event)
+            expect(result.event.id).to be_present
+            expect(result.event.event_name).to be_nil
+          end
+        end
+      end
+
+      describe "but the data required for work is invalid" do
+        describe "because the value of `event_name` is wrong" do
+          let(:event_name) { "sent" }
+
+          it "returns expected error" do
+            expect { perform }.to(
+              raise_error(
+                ApplicationService::Errors::InputError,
+                "[Usual::Example55] Wrong value in `event_name`, must be one of " \
+                "`[\"created\", \"rejected\", \"approved\"]`"
+              )
+            )
           end
         end
       end
     end
 
     context "when the input arguments are invalid" do
-      context "when `first_name`" do
-        it_behaves_like "input required check", name: :first_name
-        it_behaves_like "input type check", name: :first_name, expected_type: String
-      end
-
-      context "when `middle_name`" do
-        it_behaves_like "input type check", name: :middle_name, expected_type: String
-      end
-
-      context "when `last_name`" do
-        it_behaves_like "input required check", name: :last_name
-        it_behaves_like "input type check", name: :last_name, expected_type: String
+      context "when `event_name`" do
+        it_behaves_like "input type check", name: :event_name, expected_type: String
       end
     end
   end