+++ spec/examples/usual/example57_spec.rb
@@ -1,16 +1,21 @@
 # frozen_string_literal: true
 
-RSpec.describe Wrong::Example13 do
+RSpec.describe Usual::Example57 do
   describe ".call!" do
     subject(:perform) { described_class.call! }
 
+    include_examples "check class info",
+                     inputs: %i[],
+                     internals: %i[],
+                     outputs: %i[number]
+
     context "when the input arguments are valid" do
       describe "but the data required for work is invalid" do
-        it "returns expected error" do
+        it "returns the expected value in `first_id`" do
           expect { perform }.to(
             raise_error(
               ApplicationService::Errors::Failure,
-              "Nothing to perform. Use `make` or create a `call` method."
+              "rollback with bad number"
             )
           )
         end
@@ -21,6 +26,11 @@ RSpec.describe Wrong::Example13 do
   describe ".call" do
     subject(:perform) { described_class.call }
 
+    include_examples "check class info",
+                     inputs: %i[],
+                     internals: %i[],
+                     outputs: %i[number]
+
     context "when the input arguments are valid" do
       describe "but the data required for work is invalid" do
         include_examples "failure result class"
@@ -30,7 +40,7 @@ RSpec.describe Wrong::Example13 do
 
           expect(result.error).to be_a(ApplicationService::Errors::Failure)
           expect(result.error).to an_object_having_attributes(
-            message: "Nothing to perform. Use `make` or create a `call` method."
+            message: "rollback with bad number"
           )
         end
       end